all ring removal tests combined and updated
