+++ savu/test/travis/plugin_tests/filter_tests/ring_removal/__init__.py
