+++ test_data/test_process_lists/ring_removal/ring_removal_normalization_test.nxs
